# SINGA-490 Optimize performance of stochastic gradient descent (SGD)