--- a/examples/autograd/resnet_dist.py
+++ b/examples/autograd/resnet_dist.py
@@ -24,7 +24,6 @@ from singa import autograd
 from singa import tensor
 from singa import device
 from singa import opt
-#from singa import dist_opt
 
 import numpy as np
 from tqdm import trange
@@ -36,11 +35,10 @@ if __name__ == "__main__":
     from resnet import resnet50
     model = resnet50()
 
-    print("Start intialization............")
+    if (sgd.rank_in_global == 0):
+        print("Start intialization...........", flush=True)
 
-    #sgd = dist_opt.Dist_SGD(lr=0.1)
     dev = device.create_cuda_gpu_on(sgd.rank_in_local)
-    # dev = device.create_cuda_gpu()
     niters = 100
     batch_size = 32
     IMG_SIZE = 224
@@ -55,32 +53,39 @@ if __name__ == "__main__":
 
     import time
 
+    dev.Sync()
     start = time.time()
     fd = 0
     softmax = 0
     update = 0
     with trange(niters) as t:
         for b in t:
+            dev.Sync()
             tick = time.time()
             x = model(tx)
+            dev.Sync()
             fd += time.time() - tick
             tick = time.time()
             loss = autograd.softmax_cross_entropy(x, ty)
+            dev.Sync()
             softmax += time.time() - tick
             for p, g in autograd.backward(loss):
-                # print(p.shape, g.shape)
+                dev.Sync()  # this "for" loops for a large number of times, so can slow down
                 tick = time.time()
                 sgd.update(p, g)
+                dev.Sync()  # this "for" loops for a large number of times, so can slow down
                 update += time.time() - tick
-                # pass
 
+    dev.Sync()            
     end = time.time()
     throughput = sgd.world_size * niters * batch_size / (end - start)
-    print("Throughput = {} per second".format(throughput))
     titer = (end - start) / niters
     tforward = fd / niters
     tsoftmax = softmax / niters
     tbackward = titer - tforward - tsoftmax
     tsgd = update / niters
-    print("Total={}, forward={}, softmax={}, backward={}, sgd={}".format(
-        titer, tforward, tsoftmax, tbackward, tsgd))
+
+    if (sgd.rank_in_global == 0):
+        print("\nThroughput = {} per second".format(throughput), flush=True)
+        print("Total={}, forward={}, softmax={}, backward={}, sgd={}".format(
+        titer, tforward, tsoftmax, tbackward, tsgd), flush=True)