--- a/src/core/device/cuda_gpu.cc
+++ b/src/core/device/cuda_gpu.cc
@@ -122,5 +122,9 @@ void CudaGPU::Free(void* ptr) {
   }
 }
 
+void CudaGPU::Sync() {
+  CUDA_CHECK(cudaStreamSynchronize(ctx_.stream));
+}
+
 }  // namespace singa
 #endif  // USE_CUDA