@chrishkchris 2019-09-18
Next, I further improve the time performance by using in-place elementwise multiply in many functions such as ReLU. Here is the result:

```
ubuntu@ip-172-31-39-137:~/incubator-singa/examples/autograd$ python3 mnist_cnn.py
Starting Epoch 0:
Training loss = 585.555542, training accuracy = 0.793390
Evaluation accuracy = 0.939303, Elapsed Time = 4.206943s
Starting Epoch 1:
Training loss = 234.893921, training accuracy = 0.922409
Evaluation accuracy = 0.955729, Elapsed Time = 4.101450s
Starting Epoch 2:
Training loss = 169.515244, training accuracy = 0.943286
Evaluation accuracy = 0.970252, Elapsed Time = 4.104907s
Starting Epoch 3:
Training loss = 136.331894, training accuracy = 0.954442
Evaluation accuracy = 0.968450, Elapsed Time = 4.115959s
Starting Epoch 4:
Training loss = 118.268318, training accuracy = 0.960512
Evaluation accuracy = 0.971755, Elapsed Time = 4.117009s
Starting Epoch 5:
Training loss = 104.006439, training accuracy = 0.965732
Evaluation accuracy = 0.978866, Elapsed Time = 4.117350s
Starting Epoch 6:
Training loss = 93.860809, training accuracy = 0.969067
Evaluation accuracy = 0.977464, Elapsed Time = 4.106471s
Starting Epoch 7:
Training loss = 88.009178, training accuracy = 0.970251
Evaluation accuracy = 0.982873, Elapsed Time = 4.116037s
Starting Epoch 8:
Training loss = 81.978348, training accuracy = 0.972802
Evaluation accuracy = 0.983974, Elapsed Time = 4.121274s
Starting Epoch 9:
Training loss = 75.998878, training accuracy = 0.974103
Evaluation accuracy = 0.982272, Elapsed Time = 4.122591s
```